Move putNewStoriesUsers to stories BackendAccess
