+++ src/commons/sagas/RequestsSaga.ts
@@ -1,5 +1,4 @@
 import { call } from 'redux-saga/effects';
-import Constants from 'src/commons/utils/Constants';
 
 import {
   AchievementGoal,
@@ -15,7 +14,6 @@ import {
   WebSocketEndpointInformation
 } from '../../features/remoteExecution/RemoteExecutionTypes';
 import { PlaybackData, SourcecastData } from '../../features/sourceRecorder/SourceRecorderTypes';
-import { NameUsernameRole } from '../../pages/academy/adminPanel/subcomponents/AddStoriesUserPanel';
 import { UsernameRoleGroup } from '../../pages/academy/adminPanel/subcomponents/AddUserPanel';
 import { store } from '../../pages/createStore';
 import {
@@ -1179,20 +1177,6 @@ export const putNewUsers = async (
   provider: string
 ): Promise<Response | null> => {
   const resp = await request(`${courseId()}/admin/users`, 'PUT', {
-    ...tokens,
-    body: { users, provider },
-    noHeaderAccept: true
-  });
-
-  return resp;
-};
-
-export const putNewStoriesUsers = async (
-  tokens: Tokens,
-  users: NameUsernameRole[],
-  provider: string
-): Promise<Response | null> => {
-  const resp = await request(`${Constants.storiesBackendUrl}/stories`, 'PUT', {
     ...tokens,
     body: { users, provider },
     noHeaderAccept: true
